requirements.txt
streamlit
pinecone-client
transformers
torch
Pillow
scikit-learn
matplotlib
openai